When making http calls to save logs, do that in background
Now database logger is pretty simple: it doesn't do anything except
launching a task to put a message in a channel. This channel is read by
a background worker which actually makes all the required http calls.
This way we follow recommendation from official [docs](https://docs.microsoft.com/en-us/aspnet/core/fundamentals/logging/?view=aspnetcore-6.0#no-asynchronous-logger-methods).
It would be better to use hosted services but they don't work in blazor
wasm :(

Another issue this pr addresses is dependencies misuse. Logger providers
are of singleton scope while it's dependencies are not which is
discouraged practice of handling the dependencies. Components with
higher lifetime should not depend on components with lower lifetime scope.
To be honest that was the primary reason of change.